feat: added handle to get cadences

+++ Cadence/Models/Config/AnticipatorConfig.cs
@@ -5,6 +5,6 @@ public class AnticipatorConfig : ConfigBase
     public int ColumnSize { get; set; }
     public int MinToAnticipate { get; set; }
     public int MaxToAnticipate { get; set; }
-    public double AnticipateCadence { get; set; }
-    public double DefaultCadence { get; set; }
+    public int AnticipateCadence { get; set; }
+    public float DefaultCadence { get; set; }
 }
